agents/base-agent/requirements.txt
# Flask web framework
Flask==3.0.0
Werkzeug==3.0.1

# Azure SDKs
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
redis==5.0.1

# Application Insights and OpenTelemetry
azure-monitor-opentelemetry==1.2.0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Production server
gunicorn==21.2.0
